feat(sampling): rename semantic similarity
